fix: redis관련 의존성 추가
