--- src/main/java/com/team/leaf/user/account/dto/common/LogoutRequest.java
@@ -1,10 +1,0 @@
-package com.team.leaf.user.account.dto.common;
-
-import lombok.Data;
-
-@Data
-public class LogoutRequest {
-
-    private String email;
-
-}
