--- src/main/java/com/team/leaf/user/account/dto/request/oauth/OAuth2LoginRequest.java
@@ -1,12 +1,0 @@
-package com.team.leaf.user.account.dto.request.oauth;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class OAuth2LoginRequest {
-
-    private String email;
-
-}
